Move parallel.h to src
