+++ include/manifold/CMakeLists.txt
@@ -3,7 +3,6 @@ set(MANIFOLD_PUBLIC_HDRS
 	iters.h
 	manifold.h
 	optional_assert.h
-	parallel.h
 	polygon.h
 	vec_view.h
 	)
