+++ scripts/test-cmake.sh
@@ -12,7 +12,6 @@ EOT
 
 cat <<EOT > test.cpp
 #include <manifold/manifold.h>
-#include <manifold/parallel.h>
 int main() { manifold::Manifold foo; return 0; }
 EOT
