+++ src/CMakeLists.txt
@@ -40,6 +40,7 @@ set(MANIFOLD_PRIVATE_HDRS
 	hashtable.h
 	impl.h
 	mesh_fixes.h
+	parallel.h
 	quickhull.h
 	shared.h
 	sparse.h
