+++ src/collider.h
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 #pragma once
+#include "./parallel.h"
 #include "./sparse.h"
 #include "./utils.h"
 #include "./vec.h"
 #include "manifold/common.h"
-#include "manifold/parallel.h"
 
 #ifdef _MSC_VER
 #include <intrin.h>
