+++ src/edge_op.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "impl.h"
-#include "manifold/parallel.h"
+#include "./impl.h"
+#include "./parallel.h"
 
 namespace {
 using namespace manifold;
