+++ src/manifold.cpp
@@ -16,10 +16,10 @@
 #include <map>
 #include <numeric>
 
-#include "boolean3.h"
-#include "csg_tree.h"
-#include "impl.h"
-#include "manifold/parallel.h"
+#include "./boolean3.h"
+#include "./csg_tree.h"
+#include "./impl.h"
+#include "./parallel.h"
 
 namespace {
 using namespace manifold;
