+++ src/sdf.cpp
@@ -16,10 +16,10 @@
 
 #include "./hashtable.h"
 #include "./impl.h"
+#include "./parallel.h"
 #include "./utils.h"
 #include "./vec.h"
 #include "manifold/manifold.h"
-#include "manifold/parallel.h"
 
 namespace {
 using namespace manifold;
