+++ src/shared.h
@@ -14,10 +14,10 @@
 
 #pragma once
 
+#include "./parallel.h"
 #include "./sparse.h"
 #include "./utils.h"
 #include "./vec.h"
-#include "manifold/parallel.h"
 
 namespace manifold {
