+++ src/smoothing.cpp
@@ -15,8 +15,8 @@
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/quaternion.hpp>
 
-#include "impl.h"
-#include "manifold/parallel.h"
+#include "./impl.h"
+#include "./parallel.h"
 
 namespace {
 using namespace manifold;
