+++ src/vec.h
@@ -20,7 +20,7 @@
 #define TracyFreeS(ptr, n) (void)0
 #endif
 
-#include "manifold/parallel.h"
+#include "./parallel.h"
 #include "manifold/vec_view.h"
 
 namespace manifold {
